Refactor the project layout. (Ab)Use `attr`s.
