+++ docs/ref/saltfactories.factories.cli.rst
@@ -1,0 +1,66 @@
+saltfactories.factories.cli package
+===================================
+
+.. automodule:: saltfactories.factories.cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+saltfactories.factories.cli.call module
+---------------------------------------
+
+.. automodule:: saltfactories.factories.cli.call
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.client module
+-----------------------------------------
+
+.. automodule:: saltfactories.factories.cli.client
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.cp module
+-------------------------------------
+
+.. automodule:: saltfactories.factories.cli.cp
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.key module
+--------------------------------------
+
+.. automodule:: saltfactories.factories.cli.key
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.run module
+--------------------------------------
+
+.. automodule:: saltfactories.factories.cli.run
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.salt module
+---------------------------------------
+
+.. automodule:: saltfactories.factories.cli.salt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.cli.ssh module
+--------------------------------------
+
+.. automodule:: saltfactories.factories.cli.ssh
+   :members:
+   :undoc-members:
+   :show-inheritance:
