+++ docs/ref/saltfactories.factories.daemons.rst
@@ -1,0 +1,50 @@
+saltfactories.factories.daemons package
+=======================================
+
+.. automodule:: saltfactories.factories.daemons
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+saltfactories.factories.daemons.master module
+---------------------------------------------
+
+.. automodule:: saltfactories.factories.daemons.master
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.daemons.minion module
+---------------------------------------------
+
+.. automodule:: saltfactories.factories.daemons.minion
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.daemons.proxy module
+--------------------------------------------
+
+.. automodule:: saltfactories.factories.daemons.proxy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.daemons.sshd module
+-------------------------------------------
+
+.. automodule:: saltfactories.factories.daemons.sshd
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+saltfactories.factories.daemons.syndic module
+---------------------------------------------
+
+.. automodule:: saltfactories.factories.daemons.syndic
+   :members:
+   :undoc-members:
+   :show-inheritance:
