+++ docs/ref/saltfactories.factories.rst
@@ -6,45 +6,30 @@ saltfactories.factories package
    :undoc-members:
    :show-inheritance:
 
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   saltfactories.factories.cli
+   saltfactories.factories.daemons
+
 Submodules
 ----------
+
+saltfactories.factories.base module
+-----------------------------------
+
+.. automodule:: saltfactories.factories.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 saltfactories.factories.manager module
 --------------------------------------
 
 .. automodule:: saltfactories.factories.manager
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.factories.master module
--------------------------------------
-
-.. automodule:: saltfactories.factories.master
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.factories.minion module
--------------------------------------
-
-.. automodule:: saltfactories.factories.minion
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.factories.proxy module
-------------------------------------
-
-.. automodule:: saltfactories.factories.proxy
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.factories.syndic module
--------------------------------------
-
-.. automodule:: saltfactories.factories.syndic
    :members:
    :undoc-members:
    :show-inheritance:
