+++ docs/ref/saltfactories.rst
@@ -10,6 +10,7 @@ Subpackages
 -----------
 
 .. toctree::
+   :maxdepth: 4
 
    saltfactories.factories
    saltfactories.plugins
