+++ docs/ref/saltfactories.utils.processes.rst
@@ -9,34 +9,10 @@ saltfactories.utils.processes package
 Submodules
 ----------
 
-saltfactories.utils.processes.bases module
-------------------------------------------
-
-.. automodule:: saltfactories.utils.processes.bases
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 saltfactories.utils.processes.helpers module
 --------------------------------------------
 
 .. automodule:: saltfactories.utils.processes.helpers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.utils.processes.salts module
-------------------------------------------
-
-.. automodule:: saltfactories.utils.processes.salts
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-saltfactories.utils.processes.sshd module
------------------------------------------
-
-.. automodule:: saltfactories.utils.processes.sshd
    :members:
    :undoc-members:
    :show-inheritance:
