+++ saltfactories/factories/cli/__init__.py
@@ -1,0 +1,7 @@
+from . import call
+from . import client
+from . import cp
+from . import key
+from . import run
+from . import salt
+from . import ssh
