+++ saltfactories/factories/daemons/__init__.py
@@ -1,0 +1,5 @@
+from . import master
+from . import minion
+from . import proxy
+from . import sshd
+from . import syndic
