+++ tests/functional/factories/daemons/test_master_factory.py
@@ -1,9 +1,3 @@
-"""
-    tests.functional.factories.test_proxy_minion_factory
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-    Functional tests for the salt proxy minion factory
-"""
 import pytest
 
 from saltfactories.utils import running_username
@@ -12,15 +6,15 @@ from saltfactories.utils import running_username
 def test_hook_basic_config_defaults(testdir):
     testdir.makeconftest(
         """
-        def pytest_saltfactories_proxy_minion_configuration_defaults():
+        def pytest_saltfactories_master_configuration_defaults():
             return {'zzzz': True}
         """
     )
     p = testdir.makepyfile(
         """
         def test_basic_config_override(request, salt_factories):
-            proxy_minion_config = salt_factories.configure_proxy_minion(request, 'proxy-minion-1')
-            assert 'zzzz' in proxy_minion_config
+            master_config = salt_factories.configure_master(request, 'master-1')
+            assert 'zzzz' in master_config
         """
     )
     res = testdir.runpytest("-v")
@@ -28,24 +22,33 @@ def test_hook_basic_config_defaults(testdir):
 
 
 def test_keyword_basic_config_defaults(request, salt_factories):
-    proxy_minion_config = salt_factories.configure_proxy_minion(
-        request, "proxy-minion-1", config_defaults={"zzzz": True}
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_defaults={"zzzz": True}
     )
-    assert "zzzz" in proxy_minion_config
+    assert "zzzz" in master_config
+
+
+def test_interface_config_defaults(request, salt_factories):
+    interface = "172.17.0.1"
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_defaults={"interface": interface}
+    )
+    assert master_config["interface"] != interface
+    assert master_config["interface"] == "127.0.0.1"
 
 
 def test_hook_basic_config_overrides(testdir):
     testdir.makeconftest(
         """
-        def pytest_saltfactories_proxy_minion_configuration_overrides():
+        def pytest_saltfactories_master_configuration_overrides():
             return {'zzzz': True}
         """
     )
     p = testdir.makepyfile(
         """
         def test_basic_config_override(request, salt_factories):
-            proxy_minion_config = salt_factories.configure_proxy_minion(request, 'proxy-minion-1')
-            assert 'zzzz' in proxy_minion_config
+            master_config = salt_factories.configure_master(request, 'master-1')
+            assert 'zzzz' in master_config
         """
     )
     res = testdir.runpytest("-v")
@@ -53,28 +56,37 @@ def test_hook_basic_config_overrides(testdir):
 
 
 def test_keyword_basic_config_overrides(request, salt_factories):
-    proxy_minion_config = salt_factories.configure_proxy_minion(
-        request, "proxy-minion-1", config_overrides={"zzzz": True}
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_overrides={"zzzz": True}
     )
-    assert "zzzz" in proxy_minion_config
+    assert "zzzz" in master_config
+
+
+def test_interface_config_overrides(request, salt_factories):
+    interface = "172.17.0.1"
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_overrides={"interface": interface}
+    )
+    assert master_config["interface"] != "127.0.0.1"
+    assert master_config["interface"] == interface
 
 
 def test_hook_simple_overrides_override_defaults(testdir):
     testdir.makeconftest(
         """
-        def pytest_saltfactories_proxy_minion_configuration_defaults():
+        def pytest_saltfactories_master_configuration_defaults():
             return {'zzzz': False}
 
-        def pytest_saltfactories_proxy_minion_configuration_overrides():
+        def pytest_saltfactories_master_configuration_overrides():
             return {'zzzz': True}
         """
     )
     p = testdir.makepyfile(
         """
         def test_basic_config_override(request, salt_factories):
-            proxy_minion_config = salt_factories.configure_proxy_minion(request, 'proxy-minion-1')
-            assert 'zzzz' in proxy_minion_config
-            assert proxy_minion_config['zzzz'] is True
+            master_config = salt_factories.configure_master(request, 'master-1')
+            assert 'zzzz' in master_config
+            assert master_config['zzzz'] is True
         """
     )
     res = testdir.runpytest("-v")
@@ -82,17 +94,17 @@ def test_hook_simple_overrides_override_defaults(testdir):
 
 
 def test_keyword_simple_overrides_override_defaults(request, salt_factories):
-    proxy_minion_config = salt_factories.configure_proxy_minion(
-        request, "proxy-minion-1", config_defaults={"zzzz": False}, config_overrides={"zzzz": True}
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_defaults={"zzzz": False}, config_overrides={"zzzz": True}
     )
-    assert "zzzz" in proxy_minion_config
-    assert proxy_minion_config["zzzz"] is True
+    assert "zzzz" in master_config
+    assert master_config["zzzz"] is True
 
 
 def test_hook_nested_overrides_override_defaults(testdir):
     testdir.makeconftest(
         """
-        def pytest_saltfactories_proxy_minion_configuration_defaults():
+        def pytest_saltfactories_master_configuration_defaults():
             return {
                 'zzzz': False,
                 'user': 'foobar',
@@ -102,7 +114,7 @@ def test_hook_nested_overrides_override_defaults(testdir):
                 }
             }
 
-        def pytest_saltfactories_proxy_minion_configuration_overrides():
+        def pytest_saltfactories_master_configuration_overrides():
             return {
                 'colors': {
                     'white': True,
@@ -114,10 +126,10 @@ def test_hook_nested_overrides_override_defaults(testdir):
     p = testdir.makepyfile(
         """
         def test_basic_config_override(request, salt_factories):
-            proxy_minion_config = salt_factories.configure_proxy_minion(request, 'proxy-minion-1')
-            assert 'zzzz' in proxy_minion_config
-            assert proxy_minion_config['zzzz'] is False
-            assert proxy_minion_config['colors'] == {
+            master_config = salt_factories.configure_master(request, 'master-1')
+            assert 'zzzz' in master_config
+            assert master_config['zzzz'] is False
+            assert master_config['colors'] == {
                 'black': True,
                 'white': True,
                 'grey': False
@@ -129,9 +141,9 @@ def test_hook_nested_overrides_override_defaults(testdir):
 
 
 def test_keyword_nested_overrides_override_defaults(request, salt_factories):
-    proxy_minion_config = salt_factories.configure_proxy_minion(
+    master_config = salt_factories.configure_master(
         request,
-        "proxy-minion-1",
+        "master-1",
         config_defaults={
             "zzzz": False,
             "user": "foobar",
@@ -139,15 +151,15 @@ def test_keyword_nested_overrides_override_defaults(request, salt_factories):
         },
         config_overrides={"colors": {"white": True, "grey": False}},
     )
-    assert "zzzz" in proxy_minion_config
-    assert proxy_minion_config["zzzz"] is False
-    assert proxy_minion_config["colors"] == {"black": True, "white": True, "grey": False}
+    assert "zzzz" in master_config
+    assert master_config["zzzz"] is False
+    assert master_config["colors"] == {"black": True, "white": True, "grey": False}
 
 
 def test_nested_overrides_override_defaults(testdir):
     testdir.makeconftest(
         """
-        def pytest_saltfactories_proxy_minion_configuration_defaults():
+        def pytest_saltfactories_master_configuration_defaults():
             return {
                 'zzzz': True,
                 'user': 'foobar',
@@ -158,7 +170,7 @@ def test_nested_overrides_override_defaults(testdir):
                 }
             }
 
-        def pytest_saltfactories_proxy_minion_configuration_overrides():
+        def pytest_saltfactories_master_configuration_overrides():
             return {
                 'colors': {
                     'white': False,
@@ -171,9 +183,9 @@ def test_nested_overrides_override_defaults(testdir):
     p = testdir.makepyfile(
         """
         def test_basic_config_override(request, salt_factories):
-            proxy_minion_config = salt_factories.configure_proxy_minion(
+            master_config = salt_factories.configure_master(
                 request,
-                'proxy-minion-1',
+                'master-1',
                 config_defaults={
                     'zzzz': False,
                     'user': 'foobar',
@@ -189,9 +201,9 @@ def test_nested_overrides_override_defaults(testdir):
                     }
                 }
             )
-            assert 'zzzz' in proxy_minion_config
-            assert proxy_minion_config['zzzz'] is False
-            assert proxy_minion_config['colors'] == {
+            assert 'zzzz' in master_config
+            assert master_config['zzzz'] is False
+            assert master_config['colors'] == {
                 'black': True,
                 'white': True,
                 'grey': False,
@@ -206,10 +218,10 @@ def test_nested_overrides_override_defaults(testdir):
 def test_provide_root_dir(testdir, request, salt_factories):
     root_dir = testdir.mkdir("custom-root")
     config_defaults = {"root_dir": root_dir}
-    proxy_minion_config = salt_factories.configure_proxy_minion(
-        request, "proxy_minion-1", config_defaults=config_defaults
+    master_config = salt_factories.configure_master(
+        request, "master-1", config_defaults=config_defaults
     )
-    assert proxy_minion_config["root_dir"] == root_dir
+    assert master_config["root_dir"] == root_dir
 
 
 def configure_kwargs_ids(value):
@@ -222,14 +234,12 @@ def configure_kwargs_ids(value):
     ids=configure_kwargs_ids,
 )
 def test_provide_user(request, salt_factories, configure_kwargs):
-    proxy_minion_config = salt_factories.configure_proxy_minion(
-        request, "proxy-minion-1", **configure_kwargs
-    )
+    master_config = salt_factories.configure_master(request, "master-1", **configure_kwargs)
     if not configure_kwargs:
         # salt-factories injects the current username
-        assert proxy_minion_config["user"] is not None
-        assert proxy_minion_config["user"] == running_username()
+        assert master_config["user"] is not None
+        assert master_config["user"] == running_username()
     else:
         # salt-factories does not override the passed user value
-        assert proxy_minion_config["user"] != running_username()
-        assert proxy_minion_config["user"] == "blah"
+        assert master_config["user"] != running_username()
+        assert master_config["user"] == "blah"
