+++ tests/integration/utils/processes/test_helpers.py
@@ -9,8 +9,8 @@ import time
 import pytest
 
 from saltfactories.exceptions import ProcessNotStarted
+from saltfactories.factories.daemons.master import MasterFactory
 from saltfactories.utils import cli_scripts
-from saltfactories.utils.processes.salts import SaltMaster
 
 
 try:
@@ -47,7 +47,7 @@ def test_exit_status_unknown_user(request, salt_factories, shell_tests_salt_mast
         inject_coverage=salt_factories.inject_coverage,
         inject_sitecustomize=salt_factories.inject_sitecustomize,
     )
-    proc = SaltMaster(cli_script_name=script_path, config=shell_tests_salt_master_config)
+    proc = MasterFactory(cli_script_name=script_path, config=shell_tests_salt_master_config)
     proc.start()
     iterations = salt_factories.start_timeout
     while proc.is_alive():
