+++ tests/unit/factories/base/test_salt_factory.py
@@ -1,18 +1,13 @@
-"""
-    tests.unit.utils.processes.salts.test_salt_script_base
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-    Test Salt's base script implementation
-"""
 import json
+import os
 import sys
 from collections import OrderedDict
 from unittest import mock
 
 import pytest
 
-from saltfactories.utils.processes.bases import ProcessResult
-from saltfactories.utils.processes.salts import SaltScriptBase
+from saltfactories.factories.base import SaltCliFactory
+from saltfactories.utils.processes import ProcessResult
 
 
 @pytest.fixture
@@ -59,7 +54,7 @@ def test_default_cli_flags(minion_id, config_dir, config_file, cli_script_name):
         minion_id,
         "test.ping",
     ]
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     assert cmdline == expected
 
@@ -87,7 +82,7 @@ def test_override_log_level(minion_id, config_dir, config_file, cli_script_name,
         + flag_overrides_args
         + ["test.ping"]
     )
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     assert cmdline == expected
 
@@ -115,7 +110,7 @@ def test_override_output(minion_id, config_dir, config_file, cli_script_name, fl
         + flag_overrides_args
         + ["test.ping"]
     )
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     assert cmdline == expected
 
@@ -135,7 +130,9 @@ def test_default_cli_flags_with_timeout(minion_id, config_dir, config_file, cli_
         minion_id,
         "test.ping",
     ]
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
     # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
     proc.__cli_timeout_supported__ = True
@@ -170,21 +167,25 @@ def test_default_cli_flags_with_timeout_and_timeout_kwargs(
     ]
 
     popen_mock = mock.MagicMock()
+    popen_mock.pid = os.getpid()
     popen_mock.poll = mock.MagicMock(side_effect=[None, None, None, None, True])
     popen_mock.terminate = mock.MagicMock(return_value=ProcessResult(0, "", "", cmdline=()))
-    terminate_mock = mock.MagicMock(return_value=ProcessResult(0, "", ""))
+    terminate_mock = mock.MagicMock(return_value=ProcessResult(0, "", "", cmdline=()))
 
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     with mock.patch.object(proc, "init_terminal", popen_mock), mock.patch.object(
         proc, "terminate", terminate_mock
     ):
+        proc._terminal = popen_mock
         # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
         # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
         proc.__cli_timeout_supported__ = True
         ret = proc.run(*args, **kwargs)
         assert proc._terminal_timeout_set_explicitly is True
         assert proc._terminal_timeout == explicit_timeout
-        assert ret.cmdline == expected
+        assert popen_mock.call_args[0][0] == expected  # pylint: disable=unsubscriptable-object
 
     # To confirm behavior, let's NOT pass --timeout in args
     default_timeout = 10
@@ -204,21 +205,25 @@ def test_default_cli_flags_with_timeout_and_timeout_kwargs(
     ]
 
     popen_mock = mock.MagicMock()
+    popen_mock.pid = os.getpid()
     popen_mock.poll = mock.MagicMock(side_effect=[None, None, None, None, True])
     popen_mock.terminate = mock.MagicMock(return_value=ProcessResult(0, "", "", cmdline=()))
     terminate_mock = mock.MagicMock(return_value=ProcessResult(0, "", ""))
 
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     with mock.patch.object(proc, "init_terminal", popen_mock), mock.patch.object(
         proc, "terminate", terminate_mock
     ):
+        proc._terminal = popen_mock
         # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
         # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
         proc.__cli_timeout_supported__ = True
         ret = proc.run(*args, **kwargs)
         assert proc._terminal_timeout_set_explicitly is True
         assert proc._terminal_timeout == explicit_timeout
-        assert ret.cmdline == expected
+        assert popen_mock.call_args[0][0] == expected  # pylint: disable=unsubscriptable-object
 
     # To confirm behavior, let's NOT pass --timeout in args nor _timeout in kwargs
     default_timeout = 10
@@ -237,21 +242,25 @@ def test_default_cli_flags_with_timeout_and_timeout_kwargs(
     ]
 
     popen_mock = mock.MagicMock()
+    popen_mock.pid = os.getpid()
     popen_mock.poll = mock.MagicMock(side_effect=[None, None, None, None, True])
     popen_mock.terminate = mock.MagicMock(return_value=ProcessResult(0, "", "", cmdline=()))
     terminate_mock = mock.MagicMock(return_value=ProcessResult(0, "", ""))
 
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     with mock.patch.object(proc, "init_terminal", popen_mock), mock.patch.object(
         proc, "terminate", terminate_mock
     ):
+        proc._terminal = popen_mock
         # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
         # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
         proc.__cli_timeout_supported__ = True
         ret = proc.run(*args, **kwargs)
         assert proc._terminal_timeout_set_explicitly is False
         assert proc._terminal_timeout == default_timeout
-        assert ret.cmdline == expected
+        assert popen_mock.call_args[0][0] == expected  # pylint: disable=unsubscriptable-object
 
 
 @pytest.mark.parametrize("flag", ["-t", "--timeout", "--timeout="])
@@ -278,7 +287,9 @@ def test_override_timeout(minion_id, config_dir, config_file, cli_script_name, f
         + flag_overrides_args
         + ["test.ping",]
     )
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
     # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
     proc.__cli_timeout_supported__ = True
@@ -316,7 +327,9 @@ def test_override_timeout_bad_value(minion_id, config_dir, config_file, cli_scri
         + flag_overrides_args
         + ["test.ping",]
     )
-    proc = SaltScriptBase(cli_script_name, config=config, default_timeout=default_timeout)
+    proc = SaltCliFactory(
+        cli_script_name=cli_script_name, config=config, default_timeout=default_timeout
+    )
     # We set __cli_timeout_supported__ to True just to test. This would be an attribute set
     # at the class level for Salt CLI's that support the timeout flag, like for example, salt-run
     proc.__cli_timeout_supported__ = True
@@ -349,18 +362,19 @@ def test_override_config_dir(minion_id, config_dir, config_file, cli_script_name
         + flag_overrides_args
         + ["test.ping",]
     )
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     # We set the _terminal_timeout attribute just to test. This attribute would be set when calling
     # SaltScriptBase.run() but we don't really want to call it
     cmdline = proc.build_cmdline(*args, **kwargs)
     assert cmdline == expected
 
 
-def test_process_output(cli_script_name):
+def test_process_output(cli_script_name, config_file):
     in_stdout = '"The salt master could not be contacted. Is master running?"\n'
     in_stderr = ""
     cmdline = ["--out=json"]
-    proc = SaltScriptBase(cli_script_name)
+    config = {"conf_file": config_file}
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     stdout, stderr, json_out = proc.process_output(in_stdout, in_stderr, cmdline=cmdline)
     assert stdout == json.loads(in_stdout)
     assert stderr == in_stderr
@@ -385,7 +399,7 @@ def test_jsonify_kwargs(minion_id, config_dir, config_file, cli_script_name):
     ]
     for key, value in extra_kwargs.items():
         expected.append("{}={}".format(key, value))
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     # Function **kwargs are not orderred dictionaries on some python versions
     # let's just use sorted to make sure everything is in the output
@@ -407,7 +421,7 @@ def test_jsonify_kwargs(minion_id, config_dir, config_file, cli_script_name):
     for key, value in extra_kwargs.items():
         value = json.dumps(value)
         expected.append("{}={}".format(key, value))
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     # Function **kwargs are not orderred dictionaries on some python versions
     # let's just use sorted to make sure everything is in the output
@@ -429,7 +443,7 @@ def test_jsonify_kwargs(minion_id, config_dir, config_file, cli_script_name):
     for key, value in extra_kwargs.items():
         value = json.dumps(value)
         expected.append("{}={}".format(key, value))
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     # Function **kwargs are not orderred dictionaries on some python versions
     # let's just use sorted to make sure everything is in the output
@@ -448,7 +462,7 @@ def test_jsonify_kwargs(minion_id, config_dir, config_file, cli_script_name):
         "test.ping",
         "extra={}".format(json.dumps(extra_kwargs)),
     ]
-    proc = SaltScriptBase(cli_script_name, config=config)
+    proc = SaltCliFactory(cli_script_name=cli_script_name, config=config)
     cmdline = proc.build_cmdline(*args, **kwargs)
     # Function **kwargs are not orderred dictionaries on some python versions
     # let's just use sorted to make sure everything is in the output
