+++ tests/unit/utils/processes/test_processresult.py
@@ -1,14 +1,14 @@
 """
-tests.unit.utils.processes.bases.test_processresult
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+tests.unit.utils.processes.test_processresult
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Test saltfactories.utils.processes.bases.ProcessResult
+Test saltfactories.utils.processes.ProcessResult
 """
 import textwrap
 
 import pytest
 
-from saltfactories.utils.processes.bases import ProcessResult
+from saltfactories.utils.processes import ProcessResult
 
 
 @pytest.mark.parametrize("exitcode", [None, 1.0, -1.0, "0"])
